refactor: Add support for configurable stages in the build process
This commit modifies the `configure` script to add support for a new `--stage` argument, allowing users to specify which stages of the build process to execute. Previously, all stages were executed by default. With this change, users can now choose to skip certain stages by providing the corresponding stage number as the value for the `--stage` argument.

Co-authored-by: Antonino Natale <antonio.natale97@hotmail.com>